Product Filters and Tests

+++ Backend/Application/DTOs/Product/CreateProductDTO.cs
@@ -8,11 +8,11 @@
 
         public string Description { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public int Review { get; set; }
 
-        public CreateProductDTO(string type, string name, string description, double price, int review)
+        public CreateProductDTO(string type, string name, string description, decimal price, int review)
         {
             Type = type;
             Name = name;
